requirements.txt
Brotli==1.0.7
certifi==2020.4.5.1
chardet==3.0.4
charset-normalizer==3.4.1
click==7.1.2
contourpy==1.1.1
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_renderer==1.4.1
Flask==1.1.2
Flask-Compress==1.5.0
fonttools==4.56.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.4.7
MarkupSafe==1.1.1
matplotlib==3.7.5
narwhals==1.28.0
nest-asyncio==1.6.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.11.0.86
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.0.3
Pillow==7.1.2
plotly==6.0.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2020.1
PyYAML==6.0.2
requests==2.23.0
retrying==1.3.3
scipy==1.4.1
seaborn==0.13.2
six==1.15.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.80
ultralytics-thop==2.0.14
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.20.2
